Backend/player-service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
SQLAlchemy>=1.4,<1.5
databases==0.6.1
psycopg2-binary==2.9.7
python-jose==3.3.0
PyJWT==2.8.0
pydantic==1.10.9
gunicorn==20.1.0

